fix: Move UserDto
