+++ api-server/src/main/java/com/huyeon/superspace/domain/user/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.huyeon.superspace.web.domain.user.dto;
+package com.huyeon.superspace.domain.user.dto;
 
 import com.huyeon.superspace.domain.user.entity.User;
 import lombok.*;
@@ -7,11 +7,11 @@ import java.time.LocalDate;
 
 @Getter
 public class UserDto {
-    private String email;
-    private String name;
-    private LocalDate birthday;
+    private final String email;
+    private final String name;
+    private final LocalDate birthday;
 
-    private String expireIn;
+    private final String expireIn;
 
     public UserDto(User user) {
         email = user.getEmail();
